+++ HISTORY.rst
@@ -9,6 +9,9 @@ Release History
 **IoT Central updates**
 
 * az iot central device|device-template|api-token|diagnostic help strings updated with improved language.
+* update parsing template logic to support  DTDLV2 models.
+* remove deprecated commands  1) iot central app device-twin 2) iot central app monitor-events
+
 
 **IoT Hub updates**